+++ src/main/java/roomescape/presentation/controller/page/UserPageController.java
@@ -1,10 +1,10 @@
-package roomescape.presentation.controller;
+package roomescape.presentation.controller.page;
 
 import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.GetMapping;
 
 @Controller
-public class UserController {
+public class UserPageController {
 
     @GetMapping("/reservation")
     public String reservation() {
