+++ src/main/java/roomescape/presentation/controller/rest/ReservationController.java
@@ -1,4 +1,4 @@
-package roomescape.presentation.controller;
+package roomescape.presentation.controller.rest;
 
 import java.time.LocalDate;
 import java.util.List;
