+++ src/main/java/roomescape/presentation/controller/rest/ThemeController.java
@@ -1,4 +1,4 @@
-package roomescape.presentation.controller;
+package roomescape.presentation.controller.rest;
 
 import java.util.List;
 import org.springframework.http.HttpStatus;
